first pass on transparent windows
This adds a helper function for windowCreate that is applied when a window has the 'WINDOW_TRANSPARENT' flag.

The helper function copies the background (behind where the window will appear) to the windows buffer, so that a blitBufferToBufferTrans of the windows frm background, while actually be 'transparent' (i.e. the corners will be transparent)